moved logic from plugin.py to odm_laws_helper
