+++ ckanext/odm_laws/fanstatic/resource.config
@@ -1,3 +1,3 @@
 [main]
 
-force_top = moment.js
+<!-- force_top = moment.js -->
